+++ design/aggregate/customer_profile/customer_profile.sql
@@ -23,7 +23,7 @@ CREATE TABLE `customer_profile`
     `update_user_id`      bigint                                                        NULL     DEFAULT NULL COMMENT '更新人ID',
     `update_by`           varchar(32)                                                   NULL     DEFAULT NULL COMMENT '更新人名称',
     `update_time`         bigint                                                        NULL     DEFAULT NULL COMMENT '更新时间',
-    `deleted`             tinyint(1)                                                             DEFAULT 0 NOT NULL COMMENT '删除标识 0：未删除 id：已删除',
+    `deleted`             bigint                                                                DEFAULT 0 NOT NULL COMMENT '删除标识 0：未删除 id：已删除',
     PRIMARY KEY (`id`) USING BTREE,
     UNIQUE INDEX `idx_key_email` (`email`, `deleted`) USING BTREE,
     UNIQUE INDEX `idx_nick_name` (`nick_name`, `deleted`) USING BTREE,
